Updated - config-default+ server.csv = ssm update

+++ functions/install_ssmupdate.ps1
@@ -69,6 +69,80 @@ Function Get-UpdateSteamer {
             }
         }
         Remove-Item "$currentdir\tmp" -Recurse -Force
+        Get-UpdateSteamerConfigDefault
+    }
+}
+Function Get-UpdateSteamerConfigDefault {
+    $getlocalssm = Get-ChildItem $currentdir\config-default\ -Force
+    If ($getlocalssm) {
+        ForEach ($getlocalssm in $getlocalssm ) {  
+            $global:getlocalssmname = $getlocalssm.Name
+            If ($getlocalssmname) {
+                $githubvarcontent = Invoke-WebRequest "https://raw.githubusercontent.com/Robomikel/Steam-Server-Manager/master/config-default/$getlocalssmname" -UseBasicParsing
+                If ($githubvarcontent) {
+                    $githubvarcontent = ($githubvarcontent).Content
+                    If ($githubvarcontent) {
+                        If (Test-Path $currentdir\tmp) { } Else {
+                            New-Item  . -Name 'tmp' -ItemType Directory -InformationAction  SilentlyContinue | Out-File -Append -Encoding Default  $ssmlog
+                        }
+                        New-Item  "$currentdir\tmp\$getlocalssmname\" -Force >$null 2>&1
+                        Add-Content "$currentdir\tmp\$getlocalssmname" $githubvarcontent -InformationAction  SilentlyContinue
+                        $githubvarcontenttrim = Get-Content $currentdir\tmp\$getlocalssmname | Where-Object { $_ -notlike "" }
+                        If ($githubvarcontenttrim) {
+                            $ssmcontentlocaltrim = Get-Content $currentdir\config-default\$getlocalssmname | Where-Object { $_ -notlike "" }
+                            If ($ssmcontentlocaltrim ) {
+                                if (Compare-Object ($githubvarcontenttrim ) ($ssmcontentlocaltrim )) {
+                                    Get-Infomessage 'ssmupdates' 'update'
+                                    New-Item  "$currentdir\config-default\$getlocalssmname" -Force >$null 2>&1
+                                    Add-Content "$currentdir\config-default\$getlocalssmname" $githubvarcontent -InformationAction  SilentlyContinue
+                                } 
+                                Else {
+                                    Get-Infomessage 'nossmupdates' $true
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Remove-Item "$currentdir\tmp" -Recurse -Force
+        Get-UpdateSteamerCSV
+    }
+}
+Function Get-UpdateSteamerCSV {
+    $getlocalssm = Get-ChildItem $currentdir\data\ -Force
+    If ($getlocalssm) {
+        ForEach ($getlocalssm in $getlocalssm ) {  
+            $global:getlocalssmname = $getlocalssm.Name
+            If ($getlocalssmname) {
+                $githubvarcontent = Invoke-WebRequest "https://raw.githubusercontent.com/Robomikel/Steam-Server-Manager/master/data/$getlocalssmname" -UseBasicParsing
+                If ($githubvarcontent) {
+                    $githubvarcontent = ($githubvarcontent).Content
+                    If ($githubvarcontent) {
+                        If (Test-Path $currentdir\tmp) { } Else {
+                            New-Item  . -Name 'tmp' -ItemType Directory -InformationAction  SilentlyContinue | Out-File -Append -Encoding Default  $ssmlog
+                        }
+                        New-Item  "$currentdir\tmp\$getlocalssmname\" -Force >$null 2>&1
+                        Add-Content "$currentdir\tmp\$getlocalssmname" $githubvarcontent -InformationAction  SilentlyContinue
+                        $githubvarcontenttrim = Get-Content $currentdir\tmp\$getlocalssmname | Where-Object { $_ -notlike "" }
+                        If ($githubvarcontenttrim) {
+                            $ssmcontentlocaltrim = Get-Content $currentdir\data\$getlocalssmname | Where-Object { $_ -notlike "" }
+                            If ($ssmcontentlocaltrim ) {
+                                if (Compare-Object ($githubvarcontenttrim ) ($ssmcontentlocaltrim )) {
+                                    Get-Infomessage 'ssmupdates' 'update'
+                                    New-Item  "$currentdir\data\$getlocalssmname" -Force >$null 2>&1
+                                    Add-Content "$currentdir\data\$getlocalssmname" $githubvarcontent -InformationAction  SilentlyContinue
+                                } 
+                                Else {
+                                    Get-Infomessage 'nossmupdates' $true
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Remove-Item "$currentdir\tmp" -Recurse -Force
         Write-Information 'Press Enter to Close this session' -InformationAction Continue
         Pause  
         Stop-Process -Id $PID
